+++ src/main/java/com/github/dagwud/woodlands/game/creatures/DifficultyCache.java
@@ -8,16 +8,16 @@ import java.util.Map;
 
 public class DifficultyCache
 {
-  private final Map<Integer, Difficulty> difficulties;
+  private final Map<Double, Difficulty> difficulties;
 
   DifficultyCache(DifficultyRoot root)
   {
     this.difficulties = cacheDifficulties(root);
   }
 
-  private Map<Integer, Difficulty> cacheDifficulties(DifficultyRoot root)
+  private Map<Double, Difficulty> cacheDifficulties(DifficultyRoot root)
   {
-    Map<Integer, Difficulty> difficulties = new HashMap<>();
+    Map<Double, Difficulty> difficulties = new HashMap<>();
     if (root.difficulties != null)
     {
       for (Difficulty difficulty : root.difficulties)
@@ -28,7 +28,7 @@ public class DifficultyCache
     return difficulties;
   }
 
-  public Difficulty getDifficulty(int difficulty)
+  public Difficulty getDifficulty(double difficulty)
   {
     return difficulties.get(difficulty);
   }
